# ALF performance optimization