--- a/alf/utils/value_ops.py
+++ b/alf/utils/value_ops.py
@@ -180,8 +180,9 @@ def one_step_discounted_return(rewards, values, step_types, discounts):
                                       s=values.shape[0]))
 
     is_lasts = (step_types == StepType.LAST).to(dtype=torch.float32)
-    rets = (1 - is_lasts[:-1]) * (rewards[1:] + discounts[1:] * values[1:]) + \
-                 is_lasts[:-1] * discounts[:-1] * values[:-1]
+    discounted_values = discounts * values
+    rets = (1 - is_lasts[:-1]) * (rewards[1:] + discounted_values[1:]) + \
+                 is_lasts[:-1] * discounted_values[:-1]
     return rets.detach()