The following list the speed difference after each change. Overall speedup is about 14%. Tiny things add up. 

```
  ncalls  tottime  percall  cumtime  percall filename:lineno(function)
common.update(): use lerp_
    1991    0.848    0.000    1.482    0.001 /home/weixu/code/alf/alf/utils/common.py:175(update)
=>  1991    0.043    0.000    0.701    0.000 /home/weixu/code/alf/alf/utils/common.py:175(update)

utils.get_outer_rank => remote list()
    83766    0.168    0.000    0.200    0.000 /home/weixu/code/alf/alf/nest/utils.py:172(_get_outer_rank)
    83766    0.137    0.000    0.168    0.000 /home/weixu/code/alf/alf/nest/utils.py:172(_get_outer_rank)
remove flatten
    25901    0.027    0.000    0.290    0.000 /home/weixu/code/alf/alf/nest/utils.py:151(get_outer_rank)
    25901    0.022    0.000    0.256    0.000 /home/weixu/code/alf/alf/nest/utils.py:151(get_outer_rank)

AdamTF.step(): lerp_
    5973    0.247    0.000    2.538    0.000 /home/weixu/code/alf/alf/optimizers/adam_tf.py:83(step)
=>  5973    0.237    0.000    2.300    0.000 /home/weixu/code/alf/alf/optimizers/adam_tf.py:83(step)

RingBuffer._enqueue(): precompute circular(current_pos)
    2000    0.239    0.000    1.740    0.001 /home/weixu/code/alf/alf/utils/data_buffer.py:210(_enqueue)
=>  2000    0.246    0.000    1.469    0.001 /home/weixu/code/alf/alf/utils/data_buffer.py:210(_enqueue)

RingBuffer._enqueue(): max => clamp
    2000    0.246    0.000    1.469    0.001 /home/weixu/code/alf/alf/utils/data_buffer.py:210(_enqueue)
    2000    0.239    0.000    1.424    0.001 /home/weixu/code/alf/alf/utils/data_buffer.py:210(_enqueue)

RingBuffer._enqueue(): flattened buffer
    2000    0.239    0.000    1.424    0.001 /home/weixu/code/alf/alf/utils/data_buffer.py:210(_enqueue)
=>  2000    0.278    0.000    1.410    0.001 /home/weixu/code/alf/alf/utils/data_buffer.py:212(_enqueue)

unique under assert and -O
    2000    0.128    0.000    0.949    0.000 /home/weixu/code/alf/alf/utils/data_buffer.py:212(_enqueue)
=>  2000    0.127    0.000    0.927    0.000 /home/weixu/code/alf/alf/utils/data_buffer.py:212(_enqueue)

ReplayBuffer.get_batch(): convert_device only necessary:
    2000    0.046    0.000    1.197    0.001 /home/weixu/code/alf/alf/experience_replayers/replay_buffer.py:250(get_batch)
=>  2000    0.047    0.000    1.072    0.001 /home/weixu/code/alf/alf/experience_replayers/replay_buffer.py:250(get_batch)

DataBuffer.add_batch: clamp, one time access to current_pos, current_size
    4000    0.184    0.000    0.913    0.000 /home/weixu/code/alf/alf/utils/data_buffer.py:427(add_batch)
=>  4000    0.171    0.000    0.788    0.000 /home/weixu/code/alf/alf/utils/data_buffer.py:427(add_batch)

remove nn.Module.__getattr__
       1    0.000    0.000   28.623   28.623 /home/weixu/code/alf/alf/utils/common.py:218(run_under_record_context)
       1    0.000    0.000   28.581   28.581 /home/weixu/code/alf/alf/utils/common.py:218(run_under_record_context)
=>     1    0.000    0.000   28.137   28.137 /home/weixu/code/alf/alf/utils/common.py:218(run_under_record_context)
       1    0.000    0.000   28.034   28.034 /home/weixu/code/alf/alf/utils/common.py:218(run_under_record_context)
       1    0.000    0.000   27.955   27.955 /home/weixu/code/alf/alf/utils/common.py:218(run_under_record_context)

device_ctx.device optimization:
    13992    0.010    0.000    0.106    0.000 /home/weixu/code/alf/alf/device_ctx.py:71(__enter__)
    13992    0.008    0.000    0.092    0.000 /home/weixu/code/alf/alf/device_ctx.py:75(__exit__)
=>  13992    0.008    0.000    0.024    0.000 /home/weixu/code/alf/alf/device_ctx.py:71(__enter__)
    13992    0.006    0.000    0.006    0.000 /home/weixu/code/alf/alf/device_ctx.py:76(__exit__)

AverageEpisodicSumMetric.call: append only for non-empty last_episode_indices
    6000    0.031    0.000    1.556    0.000 /home/weixu/code/alf/alf/metrics/metrics.py:178(call)
=>  6000    0.027    0.000    0.751    0.000 /home/weixu/code/alf/alf/metrics/metrics.py:178(call)

cache initial_rollout_state
    2001    0.003    0.000    0.167    0.000 /home/weixu/code/alf/alf/algorithms/algorithm.py:613(get_initial_rollout_state)
=>  2001    0.001    0.000    0.001    0.000 /home/weixu/code/alf/alf/algorithms/algorithm.py:615(get_initial_rollout_state)

stack_nests: special handling for 1
    2000    0.002    0.000    0.473    0.000 /home/weixu/code/alf/alf/nest/utils.py:136(stack_nests)
=>  2000    0.003    0.000    0.141    0.000 /home/weixu/code/alf/alf/nest/utils.py:136(stack_nests)

element_wise_squared_loss: Do not use nn.Loss class
    5973    0.015    0.000    0.354    0.000 /home/weixu/code/alf/alf/utils/losses.py:34(element_wise_squared_loss)
    5973    0.007    0.000    0.149    0.000 /home/weixu/code/alf/alf/utils/losses.py:35(element_wise_squared_loss)

TracAlgorithm._calc_change: use get_initial_predict_state
    1991    0.043    0.000    2.069    0.001 /home/weixu/code/alf/alf/algorithms/trac_algorithm.py:188(_calc_change)
=>  1991    0.040    0.000    1.907    0.001 /home/weixu/code/alf/alf/algorithms/trac_algorithm.py:188(_calc_change)

FC optimization: use lower level pytorch API
    23928    0.024    0.000    0.829    0.000 /home/weixu/code/alf/alf/layers.py:289(forward)
=>  23928    0.048    0.000    0.732    0.000 /home/weixu/code/alf/alf/layers.py:294(forward)

TimeStep: cuda()
      1    0.000    0.000   26.377   26.377 /home/weixu/code/alf/alf/utils/common.py:218(run_under_record_context)
=>    1    0.000    0.000   25.821   25.821 /home/weixu/code/alf/alf/utils/common.py:218

one_step_discounted_return: avoid repeated computation
    3982    0.355    0.000    0.464    0.000 /home/weixu/code/alf/alf/utils/value_ops.py:162(one_step_discounted_return)
=>  3982    0.321    0.000    0.429    0.000 /home/weixu/code/alf/alf/utils/value_ops.py:162(one_step_discounted_return)

_cond() in run_under_record_context: use np.array for global_step instead of Tensor
    62084    0.558    0.000    0.563    0.000 /home/weixu/code/alf/alf/utils/common.py:240(_cond)
=>  62084    0.177    0.000    0.181    0.000 /home/weixu/code/alf/alf/utils/common.py:240(_cond)

Total speedup:
       1    0.000    0.000   29.092   29.092 /home/weixu/code/alf/alf/utils/common.py:218(run_under_record_context)
=>     1    0.000    0.000   24.088   24.088 /home/weixu/code/alf/alf/utils/common.py:218(run_under_record_context)

Actual time for complete sarsa_sac_pendulum (without using cProfile):
original 757.35user 477.83system 6:37.74elapsed 310%CPU (0avgtext+0avgdata 2329668maxresident)
after:   664.31user 466.89system 5:15.92elapsed 358%CPU (0avgtext+0avgdata 2329188maxresident)
ratio:                           79.43%
```